requirements.txt
requests==2.31.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
fastapi==0.109.0
uvicorn[standard]==0.27.0
mangum==0.17.0
pydantic==2.5.3
